Renames QuizAnswer to QuizQuestionAnswer including validator and specs
