--- spec/fixtures/valid_outcome.json
@@ -1,217 +1,0 @@
-{
-  "assetVersion": 32,
-  "assetStatus": {
-    "assetStatusId": "2",
-    "assetStatusName": "VISIBLE"
-  },
-  "organization": {
-    
-  },
-  "place": {
-    "placeGuid": "0b794135-fa1a-41b4-90da-c29a381929c1",
-    "placeName": "Article",
-    "placeDsc": "",
-    "placeUrlAdr": "",
-    "addressLine1Txt": " ",
-    "addressLine2Txt": "",
-    "cityName": " ",
-    "stateProvinceCode": " ",
-    "localityName": "",
-    "postalCode": " ",
-    "countryName": "",
-    "countryCode": "",
-    "timezone": "",
-    "timezoneOffset": null,
-    "timezoneDST": null,
-    "latitude": "",
-    "longitude": "",
-    "directionsTxt": ""
-  },
-  "evergreenParentAsset": {
-    
-  },
-  "sourceSystem": {
-    "sourceSystemName": "Active.com Articles",
-    "legacyGuid": "CA4EA0B1-7377-470D-B20D-BF6BEA23F040",
-    "affiliate": "false"
-  },
-  "assetRootAsset": {
-    
-  },
-  "assetParentAsset": {
-    "assetGuid": "dd3f0941-6255-4971-ae77-7e54508e975b"
-  },
-  "market": {
-    
-  },
-  "assetGuid": "04745751-2e92-4904-a126-09e516d9d661",
-  "assetName": "Rocky",
-  "assetDsc": "",
-  "alternateName": "Rocky",
-  "timezone": "",
-  "localTimeZoneId": "",
-  "timezoneAbb": "",
-  "timezoneName": "",
-  "currencyCd": "",
-  "localeCd": "",
-  "salesStartDate": null,
-  "salesEndDate": null,
-  "urlAdr": "http:\/\/www.active.com\/quizzes\/what-movie-should-you-run-in",
-  "detailPageTemplateId": "",
-  "preferredUrlAdr": "",
-  "logoUrlAdr": "",
-  "activityStartDate": null,
-  "activityEndDate": null,
-  "donationUrlAdr": "",
-  "teamUrlAdr": "",
-  "homePageUrlAdr": "",
-  "registrationUrlAdr": "",
-  "registrantSearchUrlAdr": "",
-  "regReqMinAge": null,
-  "regReqMaxAge": null,
-  "regReqGenderCd": "",
-  "resultsUrlAdr": "",
-  "contactName": "",
-  "contactEmailAdr": "",
-  "contactPhone": "",
-  "contactTxt": "",
-  "showContact": "false",
-  "sorId": "",
-  "sorCreateDtm": null,
-  "sorCreateUserId": "",
-  "authorName": "Melanie Clancy",
-  "childIndex": "3",
-  "publishDate": "2015-07-15T14:04:00",
-  "createdDate": "2015-07-15T23:26:40",
-  "modifiedDate": "2015-07-16T18:42:45",
-  "retryDate": "2015-07-16T14:52:42",
-  "retryCounter": "2",
-  "activityRecurrences": [
-    
-  ],
-  "assetQuantity": {
-    
-  },
-  "assetLegacyData": {
-    "assetTypeId": "CA4EA0B1-7377-470D-B20D-BF6BEA23F040",
-    "typeName": "",
-    "uploadSearchUrlAdr": "",
-    "authorName": "Melanie Clancy",
-    "onlineRegistration": "",
-    "onlineDonation": "",
-    "onlineMembership": "",
-    "onlineMembershipCostAmt": "",
-    "costAmt": "",
-    "avgUserRatingTxt": "",
-    "estParticipantNb": "",
-    "maxTeamNb": "",
-    "minGuaranteedGameNb": "",
-    "multipleStartDate": "",
-    "genderRequirementTxt": "",
-    "participationCriteriaTxt": "",
-    "userCommentTxt": "",
-    "priceExtensionTxt": "",
-    "searchWeight": "",
-    "seoUrl": "",
-    "substitutionUrl": "\/quizzes\/what-movie-should-you-run-in",
-    "trackbackUrl": "http:\/\/www.active.com\/quizzes\/what-movie-should-you-run-in",
-    "eventCategories": "",
-    "isSearchable": "true",
-    "createdDate": "2015-07-15T23:26:40",
-    "modifiedDate": "2015-07-16T18:42:03"
-  },
-  "assetTags": [
-    {
-      "tag": {
-        "tagId": "232486",
-        "tagName": "content",
-        "tagDescription": "feature"
-      }
-    }
-  ],
-  "assetAttributes": [
-    
-  ],
-  "assetPrices": [
-    
-  ],
-  "assetDescriptions": [
-    {
-      "descriptionType": {
-        "descriptionTypeId": "6",
-        "descriptionTypeName": "Standard"
-      },
-      "description": "You are an intense person with a drive to win. Being counted out as the underdog doesn\u2019t stop you from competing and, in fact, in only makes you want to win even more. You may lean on others from time to time for emotional support, but when it comes down to it, you have what it takes to get the W."
-    }
-  ],
-  "assetChannels": [
-    
-  ],
-  "assetMediaTypes": [
-    {
-      "sequence": "1",
-      "mediaType": {
-        "mediaTypeId": "327",
-        "mediaTypeName": "Features & News",
-        "mediaTypeDsc": ""
-      }
-    }
-  ],
-  "assetImages": [
-    {
-      "imageUrlAdr": "http:\/\/www.active.com\/Assets\/san-elijo-state-beach.jpg",
-      "imageName": "image2",
-      "linkUrl": "",
-      "linkTarget": "",
-      "imageCaptionTxt": "",
-      "imageType": "IMAGE"
-    }
-  ],
-  "assetTopics": [
-    
-  ],
-  "assetCategories": [
-    {
-      "sequence": "1",
-      "category": {
-        "categoryId": "307",
-        "categoryTaxonomy": "Quiz\/Outcome",
-        "categoryName": "Outcome"
-      }
-    }
-  ],
-  "assetMetaInterests": [
-    
-  ],
-  "assetInterests": [
-    
-  ],
-  "assetDeals": [
-    
-  ],
-  "socialMedia": [
-    
-  ],
-  "assetComponents": [
-    
-  ],
-  "assetServiceHostName": "as-app-qa02",
-  "componentInUrlAdr": {
-    
-  },
-  "assetSeoUrls": [
-    
-  ],
-  "overrideSeourlFlag": "false",
-  "activenetTopGraphic": {
-    
-  },
-  "evergreenAssetFlag": "false",
-  "evergreenAssets": [
-    
-  ],
-  "assetReferences": [
-    
-  ],
-  "salesStatus": "registration-unavailable"
-}
